requirements.txt
aioredis==2.0.0
aiosqlite==0.17.0
alembic==1.7.4
anyio==3.3.4
asgiref==3.4.1
async-timeout==3.0.1
click==8.0.3
databases==0.5.3
fastapi==0.70.0
h11==0.12.0
idna==3.3
importlib-metadata==4.8.1
importlib-resources==5.3.0
Mako==1.1.5
MarkupSafe==2.0.1
ormar==0.10.22
pydantic==1.8.2
sniffio==1.2.0
SQLAlchemy==1.4.25
starlette==0.16.0
typing-extensions==3.10.0.2
uvicorn==0.15.0
zipp==3.6.0
